LayoutTests/fast/dom/CSSStyleDeclaration/cssstyledeclaration-properties-descriptor-expected.txt

This tests the descriptor of CSSStyleDeclaration properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS descriptor['get'] is an instance of Function
PASS descriptor['set'] is an instance of Function
PASS descriptor['enumerable'] is true
PASS descriptor['configurable'] is true
PASS successfullyParsed is true

TEST COMPLETE
